Samar & Abrar: We now have the checkboxes working in an intuitive fashion!
When we select "Select All", all checkboxes for all the chairs are checked. After this,
if we unselect a chair, the "Select All" is unselected too. After this, if we select all the chair checkboxes,
the "Select all" is auto checked! And while all this happens, the motion commnads are sent to the chairs
which are actually selected as well as their name is listed!

To achieve this we had to exclude the jquery-mobile library but it doesn't lead to any problems on mobile interfaces!
The jquery mobile library seemed to add extra style classes as well as elements to
beautify the checkbox and lable which then could not be checked or unchecked using
the ".check" attribute of the checkbox!

All is well!